requirements.txt
Pillow
dlx
numpy  # Only for average image generation